# openapi.yaml
openapi: 3.1.0
info:
  title: Maru API
  version: 0.1.0
  description: Versioned API for the Maru convention operating platform.
paths:
...
  /api/v1/organizations/{organization_id}/editions/{edition_id}/workforce/documents/me:
    get:
      operationId: workforce_list_my_onboarding_documents
      parameters:
      - in: path
        name: edition_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: organization_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - organizations
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OnboardingDocumentRequest'
          description: ''
  /api/v1/organizations/{organization_id}/editions/{edition_id}/workforce/documents/me/{document_request_id}/upload:
    post:
      operationId: workforce_upload_my_onboarding_document
      parameters:
      - in: path
        name: document_request_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: edition_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: organization_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - organizations
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/OnboardingDocumentUpload'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/OnboardingDocumentUpload'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingDocumentRequest'
          description: ''
  /api/v1/organizations/{organization_id}/editions/{edition_id}/workforce/opportunities/{opportunity_id}/applications/me:
    post:
      operationId: workforce_submit_my_volunteer_application
      parameters:
      - in: path
        name: edition_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: opportunity_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: organization_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - organizations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VolunteerApplicationSubmit'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VolunteerApplication'
          description: ''
...
  /api/v1/public/editions/{edition_id}/volunteer-opportunities:
    get:
      operationId: workforce_list_public_volunteer_opportunities
      parameters:
      - in: path
        name: edition_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - public
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VolunteerOpportunity'
          description: ''
...
components:
  schemas:
...
    OnboardingDocumentRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        document_type_code:
          type: string
        document_type_name:
          type: string
        document_type_version:
          type: integer
        status:
          type: string
        instructions:
          type: string
        due_at:
          type:
          - string
          - 'null'
          format: date-time
        requested_at:
          type: string
          format: date-time
        submitted_at:
          type:
          - string
          - 'null'
          format: date-time
        reviewed_at:
          type:
          - string
          - 'null'
          format: date-time
        review_reason:
          type: string
        original_filename:
          type: string
        upload_available:
          type: boolean
      required:
      - document_type_code
      - document_type_name
      - document_type_version
      - due_at
      - id
      - instructions
      - original_filename
      - requested_at
      - review_reason
      - reviewed_at
      - status
      - submitted_at
      - upload_available
    OnboardingDocumentUpload:
      type: object
      properties:
        document:
          type: string
          format: uri
      required:
      - document
...
    VolunteerApplication:
      type: object
      properties:
        id:
          type: string
          format: uuid
        opportunity_id:
          type: string
          format: uuid
        status:
          type: string
        submitted_at:
          type: string
          format: date-time
      required:
      - id
      - opportunity_id
      - status
      - submitted_at
    VolunteerApplicationSubmit:
      type: object
      properties:
        motivation:
          type: string
          maxLength: 2000
      required:
      - motivation
    VolunteerOpportunity:
      type: object
      properties:
        id:
          type: string
          format: uuid
        position_code:
          type: string
        position_title:
          type: string
        department_name:
          type: string
        reports_to_title:
          type:
          - string
          - 'null'
        headline:
          type: string
        description:
          type: string
        headcount:
          type: integer
        active_assignment_count:
          type: integer
        is_filled:
          type: boolean
        accepts_applications:
          type: boolean
        applications_open_at:
          type:
          - string
          - 'null'
          format: date-time
        applications_close_at:
          type:
          - string
          - 'null'
          format: date-time
      required:
      - accepts_applications
      - active_assignment_count
      - applications_close_at
      - applications_open_at
      - department_name
      - description
      - headcount
      - headline
      - id
      - is_filled
      - position_code
      - position_title
      - reports_to_title
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid